:recycle: Moving #[system] implementation to bolt-attribute
