+++ crates/bolt-lang/attribute/system/Cargo.toml
@@ -15,3 +15,4 @@ proc-macro = true
 syn = { workspace = true, features = ["visit-mut"] }
 quote = { workspace = true }
 proc-macro2 = { workspace = true }
+bolt-attribute.workspace = true
